builder: Rename cmdline 'ops' field.
In OCaml 3.12, type inference was not as smart and could not use local
information to disambiguate same-named fields in different structures.
The easiest fix for this is just to rename the field so it doesn't
have the same name as a field in the Customize_cmdline.ops struct.

(cherry picked from commit 619905ac5dfd46fa3d14095aa97fd25ef4a4cf6f)