v2v: -i ova: Derive the name from the OVA filename.
If the guest has no <Name> element in the OVF, previously we chose
"default" as the name.  This changes that so it uses a name derived
from the basename of the OVA file instead.

For example:

  virt-v2v -i ova /path/to/myguest.ova [...]

would use "myguest" as the name (assuming no <Name> was present).

Modifies the behaviour of
commit 1ae4252c93c685cb8561b55c1231502b37212b5a.

(cherry picked from commit 98619f9c26076d034574b8822dfcc6799741ed96)