dib: handle lack of phases when reloading scripts list
Commit 8ee51ee39603d8a1117c471e6660bf605c7d6b3a re-scans for the
available scripts when running certain phases; OTOH, some of them may be
missing, usually due to the lack of scripts for that phase in the
selected set of elements.

Indeed, if there is no directory for a phase, safely raise Not_found so
the case is handled as if the phase was missing in the final_hooks
Hashtbl.