v2v: -o local: Check that UEFI firmware is installed before conversion (RHBZ#1429506).
Avoids a lengthy conversion followed by failure if we discover at the
end that OVMF is not installed.

This also changes the order of the methods in -o libvirt and -o qemu
so that it matches the order in the class interface, and also
logically makes more sense.

Thanks: Christopher Brown