configure: fix check for attribute cleanup
In cross-compilation, it is not possible to run tests, so the configure
aborts when checking for attribute cleanup.

The comment above the test hints at why we would need to do an actual
run test rather than a compile test: gcc only warns on unrecognised
attributes.

But warnings can be turned into error with -Werror.

Change the AC_RUN_IFELSE into an AC_COMPILE_IFELSE test that runs with
-Werror set. Save the CFLAGS before adding -Werror, and restore them
after the test.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>