configure: output paths where java is searched in
Output each path where we attempt to find 'java', so it is easier to
spot whether the location for the current OS is missing.  This also
removes the need to print the location next to the version, since the
location was already printed above.

(cherry picked from commit b95d9a74bb081b0f2261284638a6023180d02063)