lib: qemu: Refactor guestfs_int_test_qemu so it doesn't return qemu version.
Rather unnecessarily this function returned the parsed qemu version.
This complicates further refactoring, so I have changed the function
not to return this, and instead there is a separate function you have
to call to get the version struct (‘guestfs_int_qemu_version’).

Apart from a tiny amount of extra copying this is simply refactoring
of the interface between the direct-mode backend and the qemu query
functions.

For 1.36:
  Cherry picked from commit 0eb1380ddc9621d678661e831fad502a2830891f
  and modified in a few places because 1.36 still supports ancient
  qemu that predated virtio-scsi.