v2v: rework handling of CPU topology
Instead of storing the number of sockets, cores, and threads separately
with the possibility to have each of them optional, store them together,
so either we have a CPU topology or none.  This slightly simplifies the
handling of CPU topology.

Most of the output modes/producers already considered a missing
information of a topology as "1" when any of the other was specified, so
the behaviour is unchanged.  The only behaviour changes are when
parsing, and creating libvirt XMLs:
- libvirt requires all the three attributes to be specified, so when
  reading use them only if <topology> has all of them
- for the same reason, write <topology> only when all of them are
  specified

Adapt the vmx, and ova input modes to consider 1 threads when they have
the other information of a CPU topology, since none of them can provide
that information yet.

(cherry picked from commit ac398265c4579a985f3687d0097eb5b6494ff661)