v2v: tests: Set VIRTIO_WIN to point to phony virtio drivers directory.
As well as being consistent with other tests in this directory this
also fixes a problem when we ran the tests and the real virtio-win
package happened to be installed on the same machine.

The test used to pick up the virtio drivers from the virtio-win
package, resulting in different test output (because virtio drivers
had been installed).

Now we use the phony virtio drivers in all cases and also update the
test output (because virtio drivers are now always installed).

(cherry picked from commit 2274ee593e6c9573631e71fa4ed349615254c97f)