v2v: -i ova: Sanity check hrefs to ensure they stay inside the tarball.
Check for hrefs like "../../../../dev/sda", or dubious symlinks in the
tarball, which could escape from the tarball into the host system.

We use realpath(3) to do this.  Since this function does not work on
non-existent files, we must change the signature of get_file_ref so it
can return whether the file exists or not.