common/qemuopts: use the old pointer as realloc pointer
Call realloc() directly with the pointer to the old data, instead of
assigning it to the destination variable, and using that one.  The rest
of the code is the same, already properly checking for the results of
realloc(), so this mostly help static code analyzers.

(cherry picked from commit d48d0fc2d1d7cd10ac44d3cdd5d13e4adbec6b9e)