appliance: initialize the appliance_files struct
Some compilers do not manage to figure out that the members of it are
set only when search_appliance() in the end returns 1, which is already
checked.  Help them a bit by resetting the appliance_files struct on our
own, so they will not report that 'appliance.kernel', and the others are
used as uninitialized.

(cherry picked from commit 2dd9fd8530422c87131cd3afc97f6b15d75f4ed8)